# docs/swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: メダカライブオークション - 管理者API
  description: |
    メダカライブオークションシステムの管理者向けAPI仕様書
    
    ## 認証
    すべてのエンドポイントは`Bearer`トークン認証が必要です。
    また、`admin`ロールを持つユーザーのみアクセス可能です。
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: Development server
...
paths:
...
  /admin/announcements/{id}/toggle-visibility:
    patch:
      tags:
        - Announcements
      summary: 表示/非表示切り替え
      operationId: toggleVisibility
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 切り替え成功
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Live ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: リソースが見つかりません。
...